requirements.txt
appdirs==1.4.3
asgi-redis==1.3.0
asgiref==1.1.2
attrs==17.1.0
autobahn==17.5.1
Automat==0.6.0
channels==1.1.3
constantly==15.1.0
daphne==1.2.0
Django==1.11.1
django-core==1.4.1
django-cors-headers==2.0.2
djangorestframework==3.6.3
incremental==16.10.1
msgpack-python==0.4.8
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
redis==2.10.5
six==1.10.0
Twisted==17.1.0
txaio==2.7.1
zope.interface==4.4.1
